pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labelicious"
version = "0.1.0"
authors = [{ name = "Polat Bulut" }]
description = "LLM-powered dataset labeler"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pydantic>=2.7",
  "typer>=0.12",
  "tqdm>=4.66",
  "pandas>=2.2",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "tenacity>=9.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.2", "ruff>=0.6", "black>=24.3.0", "mypy>=1.10"]

[project.scripts]
labelicious = "labelicious.cli:app"
